+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - `csvtk join`: fix loading file with no records.
     - `csvtk filter2/muate2`:
         - support variable format of `${var}` with special charactors including commas, spaces, and parentheses, e.g., `${a,b}`, `${a b}`, or `${a (b)}`. [#186](https://github.com/shenwei356/csvtk/issues/186)
+    - `csvtk sort`: fix checking non-existed fileds.
 - [csvtk v0.24.0](https://github.com/shenwei356/csvtk/releases/tag/v0.24.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.24.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.24.0)
     - **Incompatible changes**: